Simplify the action of StoredProcedureHelper
The original way of executing a query in StoredProcedureHelper is to open a scan then find one record. However, in some cases , we need to find more than one record or check whether there is any record without throwing an exception. This was hard to write in the original design, which was misleading. So, we simplify the action of StoredProcedureHelper, which only open a scan without doing further actions.